Pull out Camera interface
